+++ tests/hardhat/Fork/RiskFund.ts
@@ -484,15 +484,26 @@ describe("Risk Fund: Tests", function () {
     });
 
     describe("swapPoolsAssets", async function () {
+      let deadline: number;
+
+      beforeEach(async function () {
+        deadline = (await ethers.provider.getBlock("latest")).timestamp + 100;
+      });
+
       it("fails if called with incorrect arguments", async function () {
         await expect(
-          riskFund.swapPoolsAssets([vUSDT.address, vUSDC.address], [convertToUnit(10, 18)], []),
+          riskFund.swapPoolsAssets([vUSDT.address, vUSDC.address], [convertToUnit(10, 18)], [], deadline),
         ).to.be.revertedWith("Risk fund: markets and amountsOutMin are unequal lengths");
       });
 
       it("fails if called with incorrect path length", async function () {
         await expect(
-          riskFund.swapPoolsAssets([vUSDT.address, vUSDC.address], [convertToUnit(10, 18), convertToUnit(10, 18)], []),
+          riskFund.swapPoolsAssets(
+            [vUSDT.address, vUSDC.address],
+            [convertToUnit(10, 18), convertToUnit(10, 18)],
+            [],
+            deadline,
+          ),
         ).to.be.revertedWith("Risk fund: markets and paths are unequal lengths");
       });
 
@@ -502,7 +513,7 @@ describe("Risk Fund: Tests", function () {
         await vUSDC.reduceReserves(convertToUnit(100, 18));
         await protocolShareReserve.releaseFunds(comptroller1Proxy.address, USDC.address, convertToUnit(100, 18), 0);
         await expect(
-          riskFund.swapPoolsAssets([vUSDC.address], [convertToUnit(10, 18)], [[USDT.address, BUSD.address]]),
+          riskFund.swapPoolsAssets([vUSDC.address], [convertToUnit(10, 18)], [[USDT.address, BUSD.address]], deadline),
         ).to.be.revertedWith("RiskFund: swap path must start with the underlying asset");
       });
 
@@ -512,7 +523,7 @@ describe("Risk Fund: Tests", function () {
         await vUSDC.reduceReserves(convertToUnit(100, 18));
         await protocolShareReserve.releaseFunds(comptroller1Proxy.address, USDC.address, convertToUnit(100, 18), 0);
         await expect(
-          riskFund.swapPoolsAssets([vUSDC.address], [convertToUnit(10, 18)], [[USDC.address, USDT.address]]),
+          riskFund.swapPoolsAssets([vUSDC.address], [convertToUnit(10, 18)], [[USDC.address, USDT.address]], deadline),
         ).to.be.revertedWith("RiskFund: finally path must be convertible base asset");
       });
 
@@ -539,6 +550,7 @@ describe("Risk Fund: Tests", function () {
               [USDT.address, BUSD.address],
               [USDC.address, BUSD.address],
             ],
+            deadline,
           ),
         ).to.be.revertedWithCustomError(misconfiguredRiskFund, "ZeroAddressNotAllowed");
       });
@@ -567,6 +579,12 @@ describe("Risk Fund: Tests", function () {
   });
 
   describe("Risk fund transfers", async function () {
+    let deadline: number;
+
+    beforeEach(async function () {
+      deadline = (await ethers.provider.getBlock("latest")).timestamp + 100;
+    });
+
     it("Checks access control", async function () {
       const [admin] = await ethers.getSigners();
       // Revoke
@@ -576,7 +594,10 @@ describe("Risk Fund: Tests", function () {
         admin.address,
       );
       // Fails
-      await expect(riskFund.swapPoolsAssets([], [], [])).to.be.revertedWithCustomError(riskFund, "Unauthorized");
+      await expect(riskFund.swapPoolsAssets([], [], [], deadline)).to.be.revertedWithCustomError(
+        riskFund,
+        "Unauthorized",
+      );
 
       // Reset
       await accessControlManager.giveCallPermission(
@@ -585,7 +606,7 @@ describe("Risk Fund: Tests", function () {
         admin.address,
       );
       // Succeeds
-      await riskFund.swapPoolsAssets([], [], []);
+      await riskFund.swapPoolsAssets([], [], [], deadline);
     });
 
     it("Convert to BUSD without funds", async function () {
@@ -605,6 +626,7 @@ describe("Risk Fund: Tests", function () {
           [USDC.address, BUSD.address],
           [USDT.address, BUSD.address],
         ],
+        deadline,
       );
       expect(amount).equal("0");
     });
@@ -616,7 +638,7 @@ describe("Risk Fund: Tests", function () {
 
       const protocolReserveUSDCBal = await USDC.balanceOf(protocolShareReserve.address);
       expect(protocolReserveUSDCBal).equal(convertToUnit(50, 18));
-      await protocolShareReserve.releaseFunds(comptroller1Proxy.address, USDC.address, convertToUnit(30, 18), 0);
+      await protocolShareReserve.releaseFunds(comptroller1Proxy.address, USDC.address, convertToUnit(18, 18), 0);
       const riskFundUSDCBal = await USDC.balanceOf(riskFund.address);
       expect(riskFundUSDCBal).equal(convertToUnit(9, 18));
 
@@ -636,6 +658,7 @@ describe("Risk Fund: Tests", function () {
           [USDC.address, BUSD.address],
           [USDT.address, BUSD.address],
         ],
+        deadline,
       );
       expect(amount).equal("0");
     });
@@ -647,7 +670,7 @@ describe("Risk Fund: Tests", function () {
       const protocolReserveUSDCBal = await USDC.balanceOf(protocolShareReserve.address);
       expect(protocolReserveUSDCBal).equal(convertToUnit(100, 18));
 
-      await protocolShareReserve.releaseFunds(comptroller1Proxy.address, USDC.address, convertToUnit(100, 18), 0);
+      await protocolShareReserve.releaseFunds(comptroller1Proxy.address, USDC.address, convertToUnit(60, 18), 0);
       const riskFundUSDCBal = await USDC.balanceOf(riskFund.address);
       expect(riskFundUSDCBal).equal(convertToUnit(30, 18));
 
@@ -667,6 +690,7 @@ describe("Risk Fund: Tests", function () {
           [USDC.address, BUSD.address],
           [USDT.address, BUSD.address],
         ],
+        deadline,
       );
 
       expect(await riskFund.getPoolsBaseAssetReserves(comptroller1Proxy.address)).to.be.equal("29916047622748892393");
@@ -697,8 +721,8 @@ describe("Risk Fund: Tests", function () {
       const protocolReserveUSDTBal = await USDT.balanceOf(protocolShareReserve.address);
       expect(protocolReserveUSDTBal).equal(convertToUnit(100, 18));
 
-      await protocolShareReserve.releaseFunds(comptroller1Proxy.address, USDC.address, convertToUnit(100, 18), 0);
-      await protocolShareReserve.releaseFunds(comptroller1Proxy.address, USDT.address, convertToUnit(100, 18), 0);
+      await protocolShareReserve.releaseFunds(comptroller1Proxy.address, USDC.address, convertToUnit(60, 18), 0);
+      await protocolShareReserve.releaseFunds(comptroller1Proxy.address, USDT.address, convertToUnit(60, 18), 0);
       const riskFundUSDCBal = await USDC.balanceOf(riskFund.address);
       expect(riskFundUSDCBal).equal(convertToUnit(30, 18));
 
@@ -721,6 +745,7 @@ describe("Risk Fund: Tests", function () {
           [USDC.address, BUSD.address],
           [USDT.address, BUSD.address],
         ],
+        deadline,
       );
 
       expect(await riskFund.getPoolsBaseAssetReserves(comptroller1Proxy.address)).to.be.equal("59832095245497784786");
@@ -734,6 +759,12 @@ describe("Risk Fund: Tests", function () {
   });
   // myContract.connect(myFake.wallet).doSomething();
   describe("Transfer to Auction contract", async function () {
+    let deadline: number;
+
+    beforeEach(async function () {
+      deadline = (await ethers.provider.getBlock("latest")).timestamp + 100;
+    });
+
     it("Revert while transfering funds to Auction contract", async function () {
       await expect(
         riskFund.connect(busdUser).transferReserveForAuction(comptroller1Proxy.address, convertToUnit(30, 18)),
@@ -784,6 +815,7 @@ describe("Risk Fund: Tests", function () {
           [USDC.address, BUSD.address],
           [USDT.address, BUSD.address],
         ],
+        deadline,
       );
 
       const beforeTransfer = await BUSD.balanceOf(shortfall.address);
@@ -831,6 +863,7 @@ describe("Risk Fund: Tests", function () {
           [USDC.address, BUSD.address],
           [USDT.address, BUSD.address],
         ],
+        deadline,
       );
 
       expect(await riskFund.getPoolsBaseAssetReserves(comptroller1Proxy.address)).to.be.equal(0);
@@ -938,12 +971,12 @@ describe("Risk Fund: Tests", function () {
       let riskUSDTFor3 = await riskFund.getPoolAssetReserve(comptroller3Proxy.address, USDT.address);
       let riskBUSDTFor3 = await riskFund.getPoolAssetReserve(comptroller3Proxy.address, BUSD.address);
 
-      expect(riskUSDTFor1).equal(convertToUnit(30, 18));
-      expect(riskUSDCFor1).equal(convertToUnit(27, 18));
-      expect(riskUSDTFor2).equal(convertToUnit(33, 18));
-      expect(riskUSDCFor2).equal(convertToUnit(24, 18));
-      expect(riskUSDTFor3).equal(convertToUnit(39, 18));
-      expect(riskBUSDTFor3).equal(convertToUnit(15, 18));
+      expect(riskUSDTFor1).equal(convertToUnit(50, 18));
+      expect(riskUSDCFor1).equal(convertToUnit(45, 18));
+      expect(riskUSDTFor2).equal(convertToUnit(55, 18));
+      expect(riskUSDCFor2).equal(convertToUnit(40, 18));
+      expect(riskUSDTFor3).equal(convertToUnit(65, 18));
+      expect(riskBUSDTFor3).equal(convertToUnit(25, 18));
 
       await riskFund.swapPoolsAssets(
         [vUSDT.address, vUSDC.address, vUSDT2.address, vUSDC2.address, vUSDT3.address, vBUSD3.address],
@@ -963,9 +996,10 @@ describe("Risk Fund: Tests", function () {
           [USDT.address, BUSD.address],
           [USDT.address, BUSD.address],
         ],
+        deadline,
       );
 
-      expect(await riskFund.getPoolsBaseAssetReserves(comptroller1Proxy.address)).to.be.equal("56841295980235012443");
+      expect(await riskFund.getPoolsBaseAssetReserves(comptroller1Proxy.address)).to.be.equal("94717497407756046533");
 
       riskUSDTFor1 = await riskFund.getPoolAssetReserve(comptroller1Proxy.address, USDT.address);
       riskUSDCFor1 = await riskFund.getPoolAssetReserve(comptroller1Proxy.address, USDC.address);
@@ -981,7 +1015,7 @@ describe("Risk Fund: Tests", function () {
       expect(riskUSDTFor3).equal(0);
 
       // As BUSD is base asset so PoolAssetReserves should be present.
-      expect(riskBUSDTFor3).to.be.closeTo(convertToUnit(53, 18), convertToUnit(9, 17));
+      expect(riskBUSDTFor3).to.be.closeTo(convertToUnit(90, 18), convertToUnit(9, 17));
 
       const poolReserve1 = await riskFund.getPoolsBaseAssetReserves(comptroller1Proxy.address);
 
@@ -989,9 +1023,9 @@ describe("Risk Fund: Tests", function () {
 
       const poolReserve3 = await riskFund.getPoolsBaseAssetReserves(comptroller3Proxy.address);
 
-      expect(poolReserve1).to.be.closeTo(convertToUnit(56, 18), convertToUnit(9, 17));
-      expect(poolReserve2).to.be.closeTo(convertToUnit(56, 18), convertToUnit(9, 17));
-      expect(poolReserve3).to.be.closeTo(convertToUnit(53, 18), convertToUnit(9, 17));
+      expect(poolReserve1).to.be.closeTo(convertToUnit(95, 18), convertToUnit(9, 17));
+      expect(poolReserve2).to.be.closeTo(convertToUnit(95, 18), convertToUnit(9, 17));
+      expect(poolReserve3).to.be.closeTo(convertToUnit(90, 18), convertToUnit(9, 17));
     });
   });
 });